[app_scheduler] rename library script
